requirements.txt
apache-airflow==2.7.1
dbt-postgres==1.7.3
pandas==1.5.3
numpy==1.24.3
scipy==1.10.1
psycopg2-binary==2.9.7 